+++ plugins/openchoreo/src/components/AccessControl/RolesTab/RolesTable.tsx
@@ -25,11 +25,11 @@ import {
   ListItemText,
   CircularProgress,
 } from '@material-ui/core';
-import { makeStyles } from '@material-ui/core/styles';
 import DeleteIcon from '@material-ui/icons/DeleteOutlined';
 import EditIcon from '@material-ui/icons/EditOutlined';
 import SearchIcon from '@material-ui/icons/Search';
 import WarningIcon from '@material-ui/icons/Warning';
+import { useStyles } from './styles';
 
 interface RoleRow {
   name: string;
@@ -55,50 +55,6 @@ interface RolesTableProps {
   onCheckBindings: (name: string) => Promise<BindingSummary[]>;
   onForceDelete: (name: string, bindings: BindingSummary[]) => Promise<void>;
 }
-
-const useStyles = makeStyles(theme => ({
-  searchField: {
-    width: 300,
-    marginBottom: theme.spacing(2),
-  },
-  tableContainer: {
-    marginTop: theme.spacing(2),
-  },
-  actionsChip: {
-    margin: theme.spacing(0.25),
-    maxWidth: 150,
-  },
-  actionsCell: {
-    maxWidth: 400,
-  },
-  emptyState: {
-    textAlign: 'center',
-    padding: theme.spacing(4),
-  },
-  deleteButton: {
-    borderColor: theme.palette.error.main,
-    color: theme.palette.error.main,
-    '&:hover': {
-      borderColor: theme.palette.error.dark,
-      backgroundColor: 'rgba(211, 47, 47, 0.04)',
-    },
-  },
-  warningHeader: {
-    display: 'flex',
-    alignItems: 'center',
-    gap: theme.spacing(1),
-  },
-  warningIcon: {
-    color: theme.palette.warning.main,
-  },
-  mappingsList: {
-    maxHeight: 200,
-    overflow: 'auto',
-    backgroundColor: theme.palette.background.default,
-    borderRadius: theme.shape.borderRadius,
-    marginTop: theme.spacing(2),
-  },
-}));
 
 export const RolesTable = ({
   roles,
